+++ include/aquarius/core/elastic.hpp
