+++ test/unit_test.vcxproj.filters
@@ -25,9 +25,6 @@
     <Filter Include="include\aquarius\message">
       <UniqueIdentifier>{0d5dd5db-33d3-4f0e-8872-6db7b9f0ecb3}</UniqueIdentifier>
     </Filter>
-    <Filter Include="include\aquarius\system">
-      <UniqueIdentifier>{77e47d6a-d4e0-4016-b765-e7e5d8f33e56}</UniqueIdentifier>
-    </Filter>
     <Filter Include="include\aquarius\session">
       <UniqueIdentifier>{7db6ff92-099d-452e-be84-d69c36512274}</UniqueIdentifier>
     </Filter>
@@ -48,6 +45,9 @@
     </Filter>
     <Filter Include="include\aquarius\core\impl">
       <UniqueIdentifier>{8cb9a4dc-739c-433e-9790-15f7f383bb74}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="include\aquarius\context\impl">
+      <UniqueIdentifier>{e78b2f64-a8f4-4bf6-ad81-7718c2227fab}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -81,9 +81,6 @@
       <Filter>include\aquarius</Filter>
     </ClInclude>
     <ClInclude Include="..\include\aquarius\context.hpp">
-      <Filter>include\aquarius</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\elastic.hpp">
       <Filter>include\aquarius</Filter>
     </ClInclude>
     <ClInclude Include="..\include\aquarius\io_service_pool.hpp">
@@ -127,12 +124,6 @@
     </ClInclude>
     <ClInclude Include="error_code.h">
       <Filter>头文件</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\event.hpp">
-      <Filter>include\aquarius\system</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\visitor.hpp">
-      <Filter>include\aquarius\system</Filter>
     </ClInclude>
     <ClInclude Include="..\include\aquarius\service.hpp">
       <Filter>include\aquarius</Filter>
@@ -232,6 +223,15 @@
     </ClInclude>
     <ClInclude Include="..\include\aquarius\core\impl\system_error_message.hpp">
       <Filter>include\aquarius\core\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\message\impl\visitable.hpp">
+      <Filter>include\aquarius\message\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\context\impl\visitor.hpp">
+      <Filter>include\aquarius\context\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\elastic.hpp">
+      <Filter>include\aquarius\core</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
